Notify: fix crash (indirection of null shared pointer), caused by races
Private method notify_widgets() used to modify map by
  map_[X][Y]....

 (here X is object key, Y is WApplication*)

So, if key Y is not present, the above creates it and empty pointer
to "poster" as well. For this to happen, last widget of given key
is to be removed after .emit() but prior to notify_widgets().

Currently notify_widgets() checks if needed keys are in map.

close #111